Move attributes to new file.
Change defaultMaxWidth logic again.
